% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.rna.mat}
\alias{pbmc.rna.mat}
\title{Single cell RNA-sequencing (scRNA-seq) peripheral blood (PBMC) data sample.}
\format{
A scRNA-seq data of dgCMatrix class with 1000 rows and 33538 columns
\describe{
  \item{pbmc.rna.mat}{RNA expression data for 1000 cells and 33538 genes}
}
}
\usage{
pbmc.rna.mat
}
\description{
Single cell RNA-sequencing (scRNA-seq) subset of the Hao et al. 2021 human peripheral blood mononuclear cell (PBMC) data (GEO: GSE164378, DOI: 10.1016/j.cell.2021.04.048).
}
\keyword{datasets}
