% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECK-package.R
\docType{package}
\name{SPECK-package}
\alias{SPECK-package}
\alias{_PACKAGE}
\title{SPECK: Receptor Abundance Estimation using Reduced Rank Reconstruction and Clustered Thresholding}
\description{
Surface Protein abundance Estimation using CKmeans-based clustered thresholding ('SPECK') is an unsupervised learning-based method that performs receptor abundance estimation for single cell RNA-sequencing data based on reduced rank reconstruction (RRR) and a clustered thresholding mechanism. Seurat's normalization method is described in: Hao et al., (2021) \doi{10.1016/j.cell.2021.04.048}, Stuart et al., (2019) \doi{10.1016/j.cell.2019.05.031}, Butler et al., (2018) \doi{10.1038/nbt.4096} and Satija et al., (2015) \doi{10.1038/nbt.3192}. Method for the RRR is further detailed in: Erichson et al., (2019) \doi{10.18637/jss.v089.i11} and Halko et al., (2009) \href{https://arxiv.org/abs/0909.4061}{arXiv:0909.4061}. Clustering method is outlined in: Song et al., (2020) \doi{10.1093/bioinformatics/btaa613} and Wang et al., (2011) \doi{10.32614/RJ-2011-015}.
}
\author{
\strong{Maintainer}: Azka Javaid \email{azka.javaid.gr@dartmouth.edu}

Authors:
\itemize{
  \item H. Robert Frost \email{hildreth.r.frost@dartmouth.edu}
}

}
\keyword{internal}
